requirements.txt
Flask==0.12.2
Flask-SocketIO==2.9.1
gunicorn==19.7.1
MarkupSafe==1.0
Pillow==4.2.1
opencv-python==3.2.0.7
Werkzeug==0.12.2
python-socketio==1.7.7
python-engineio==3.12.1
six==1.10.0
Jinja2==2.9.6
itsdangerous==0.24
gevent-websocket==0.10.1
eventlet==0.25.2
Flask-RESTful==0.3.8
Pyrebase==3.0.27
azure-cognitiveservices-speech==1.12.0
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dialogflow==1.0.0
google-api-core==1.17.0
google-auth==1.15.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
numpy==1.18.4
opencv-contrib-python==4.2.0.34
playsound==1.2.2
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9